requirements_render.txt
# Render-optimized requirements for CAD Analyzer Pro
streamlit==1.46.1
plotly==6.2.0
pandas==2.3.1
numpy==2.2.6
shapely==2.1.1
scipy==1.16.0
scikit-learn==1.7.0
ezdxf==1.4.2
PyMuPDF==1.26.3
opencv-python-headless==4.12.0.88
pillow==11.3.0
psycopg2-binary==2.9.10
sqlalchemy==2.0.41
python-dotenv==1.1.1
reportlab==4.4.2
matplotlib==3.10.3
networkx==3.5
psutil==7.0.0
redis==6.2.0
websockets==15.0.1
asyncio==3.4.3
requests==2.32.4
gunicorn==23.0.0